shared features of edit & rule panel in base class
